+++ browser/data-browser/src/routes/NewResource/NewRoute.tsx
@@ -49,11 +49,13 @@ function NewResourceSelector() {
   }
 
   const onUploadComplete = useCallback(
-    (files: string[]) => {
-      toast.success(`Uploaded ${files.length} files.`);
+    (fileSubjects: string[]) => {
+      toast.success(`Uploaded ${fileSubjects.length} files.`);
 
-      if (calculatedParent) {
-        navigate(constructOpenURL(calculatedParent));
+      if (fileSubjects.length > 1 && parentSubject) {
+        navigate(constructOpenURL(parentSubject));
+      } else {
+        navigate(constructOpenURL(fileSubjects[0]));
       }
     },
     [parentSubject, navigate],