+++ browser/lib/src/store.ts
@@ -2,6 +2,7 @@ import { ulid } from 'ulidx';
 import type { Agent } from './agent.js';
 import {
   removeCookieAuthentication,
+  serverSupportsRegister,
   setCookieAuthentication,
 } from './authentication.js';
 import { Client, type FileOrFileLike } from './client.js';
@@ -92,6 +93,10 @@ type StoreEventHandlers = {
 
 /** Returns True if the client has WebSocket support */
 const supportsWebSockets = () => typeof WebSocket !== 'undefined';
+
+export type ServerSupports = {
+  emailRegister: boolean;
+};
 
 /**
  * An in memory store that has a bunch of usefful methods for retrieving Atomic
@@ -736,6 +741,13 @@ export class Store {
     } else {
       console.warn('WebSockets not supported, no window available');
     }
+  }
+
+  /** Checks which features the current Server instance supports */
+  public async getServerSupports(): Promise<ServerSupports> {
+    return {
+      emailRegister: await serverSupportsRegister(this),
+    };
   }
 
   /**