+++ browser/react/src/index.ts
@@ -32,4 +32,5 @@ export * from './useServerSearch.js';
 export * from './useCollection.js';
 export * from './useMemberFromCollection.js';
 export * from './useCollectionPage.js';
+export * from './useServerSupports.js';
 export * from '@tomic/lib';